+++ libraries/core/src/topics.rs
@@ -106,53 +106,6 @@ impl DataflowResult {
     pub fn is_ok(&self) -> bool {
         self.node_results.values().all(|r| r.is_ok())
     }
-
-    pub fn root_error(&self) -> RootError<'_> {
-        RootError(self)
-    }
-}
-
-pub struct RootError<'a>(&'a DataflowResult);
-
-impl std::fmt::Display for RootError<'_> {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        let failed = self
-            .0
-            .node_results
-            .iter()
-            .filter_map(|(id, r)| r.as_ref().err().map(|e| (id, e)));
-        let total_failed = failed.clone().count();
-
-        let mut non_cascading: Vec<_> = failed
-            .clone()
-            .filter(|(_, e)| !matches!(e.cause, NodeErrorCause::Cascading { .. }))
-            .collect();
-        non_cascading.sort_by_key(|(_, e)| e.timestamp);
-        // try to print earliest non-cascading error
-        if let Some((id, err)) = non_cascading.first() {
-            write!(f, "Node `{id}` failed: {err}")?;
-        } else {
-            // no non-cascading errors -> print earliest cascading
-            let mut all: Vec<_> = failed.collect();
-            all.sort_by_key(|(_, e)| e.timestamp);
-            if let Some((id, err)) = all.first() {
-                write!(f, "Node `{id}` failed: {err}")?;
-            } else {
-                write!(f, "unknown error")?;
-            }
-        }
-
-        if total_failed > 1 {
-            write!(
-                f,
-                "\n\nThere are {} more errors. Check the `out/{}` folder for full details.",
-                total_failed - 1,
-                self.0.uuid
-            )?;
-        }
-
-        Ok(())
-    }
 }
 
 #[derive(Debug, Clone, serde::Deserialize, serde::Serialize)]
